+++ src/main/java/me/happy/rustbeds/RustBeds.java
@@ -6,6 +6,7 @@ import me.happy.rustbeds.commands.RespawnMenuCommand;
 import me.happy.rustbeds.listeners.*;
 import me.happy.rustbeds.metrics.MetricsBootstrap;
 import me.happy.rustbeds.models.BedData;
+import me.happy.rustbeds.utils.AuditLog;
 import me.happy.rustbeds.utils.PlayerBedStore;
 import me.happy.rustbeds.utils.RespawnAnchorStore;
 
@@ -29,7 +30,7 @@ import java.util.Set;
 import java.util.stream.Stream;
 
 public final class RustBeds extends JavaPlugin {
-    private static final int CURRENT_CONFIG_VERSION = 5;
+    private static final int CURRENT_CONFIG_VERSION = 9;
     private static final String CONFIG_FILE_NAME = "config.yml";
     private static final String LEGACY_PLUGIN_FOLDER_NAME = "MultipleBedSpawn";
     private static final String DATABASE_FILE_NAME = "respawn-points.db";
@@ -49,7 +50,8 @@ public final class RustBeds extends JavaPlugin {
     );
     private static final Set<String> REMOVED_CONFIG_PATHS = Set.of(
             "disable-sleeping",
-            "remove-beds-gui"
+            "remove-beds-gui",
+            "audit-log.verbose"
     );
 
     public static final String ADMIN_PERMISSION = "rustbeds.admin";
@@ -75,6 +77,7 @@ public final class RustBeds extends JavaPlugin {
         MetricsBootstrap.start(this);
         playerBedStore = new PlayerBedStore(this);
         respawnAnchorStore = new RespawnAnchorStore(this);
+        AuditLog.initialize(this);
 
         getServer().getPluginManager().registerEvents(new PlayerRespawnListener(this), this);
         getServer().getPluginManager().registerEvents(new RespawnMenuHandler(this), this);
@@ -195,6 +198,7 @@ public final class RustBeds extends JavaPlugin {
         createLanguageConfig();
         playerBedStore.reload();
         respawnAnchorStore.reload();
+        AuditLog.initialize(this);
     }
 
     @Override